requirement.txt
networkx                  3.2
numpy                     1.26.4
ogb                       1.3.6
pandas                    1.4.2
PyGCL                     0.1.2
PyYAML                    6.0.1
scikit-learn              1.4.2
scipy                     1.13.0
tensorboardX              2.6.2.2
torch                     2.1.0
torch-cluster             1.6.3+pt21cu121
torch_geometric           2.5.3
torch-scatter             2.1.2+pt21cu121
torch-sparse              0.6.18+pt21cu121
torch-spline-conv         1.2.2+pt21cu121
tqdm                      4.66.2
